Adding unique user salts
Tests failing on removing user using clear text password

+++ appauth/auth.go
@@ -3,6 +3,8 @@ package appauth
 import (
 	"encoding/hex"
 	"errors"
+	"fmt"
+	"math/rand"
 	"time"
 
 	"gopkg.in/redis.v3"
@@ -15,6 +17,7 @@ import (
 const (
 	TOKEN_TTL           = time.Hour // One hour
 	MIN_PASSWORD_LENGTH = 8
+	LETTER_BYTES        = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
 )
 
 var Config configuration.Configuration
@@ -99,17 +102,25 @@ func CreateUserPassword(user string, password string) (result string, err error)
 		return "", errors.New("appauth.CreateUserPassword: Password must be at least " + string(MIN_PASSWORD_LENGTH) + " characters")
 	}
 
-	// Generate hash
-	output, err := argon2.Key([]byte(password), []byte(Config.PasswordSalt), 3, 4, 4096, 32, argon2.Argon2i)
+	// Generate salt
+	randomStrIn := RandStringBytes(32)
+	saltOutput, err := argon2.Key([]byte(randomStrIn), []byte(Config.PasswordSalt), 3, 4, 4096, 64, argon2.Argon2i)
 	if err != nil {
 		return "", errors.New("appauth.CreateUserPassword: Could not generate secure hash. " + err.Error())
 	}
+	userSalt := hex.EncodeToString(saltOutput)
 
-	hash := hex.EncodeToString(output)
+	// Generate hash
+	userPasswordSalt := userSalt + password
+	hashOutput, err := argon2.Key([]byte(userPasswordSalt), []byte(Config.PasswordSalt), 3, 4, 4096, 64, argon2.Argon2i)
+	if err != nil {
+		return "", errors.New("appauth.CreateUserPassword: Could not generate secure hash. " + err.Error())
+	}
+	hash := hex.EncodeToString(hashOutput)
 
 	// Prepare statement for inserting data
-	insertStatement := "INSERT INTO accounts_auth (`accountNumber`, `password`, `timestamp`) "
-	insertStatement += "VALUES(?, ?, ?)"
+	insertStatement := "INSERT INTO accounts_auth (`accountNumber`, `password`, `salt`, `timestamp`) "
+	insertStatement += "VALUES(?, ?, ?, ?)"
 	stmtIns, err := Config.Db.Prepare(insertStatement)
 	if err != nil {
 		return "", errors.New("appauth.CreateUserPassword: Error with insert. " + err.Error())
@@ -120,7 +131,7 @@ func CreateUserPassword(user string, password string) (result string, err error)
 	t := time.Now()
 	sqlTime := int32(t.Unix())
 
-	_, err = stmtIns.Exec(user, hash, sqlTime)
+	_, err = stmtIns.Exec(user, hash, userSalt, sqlTime)
 
 	if err != nil {
 		return "", errors.New("appauth.CreateUserPassword: Could not save account. " + err.Error())
@@ -130,7 +141,7 @@ func CreateUserPassword(user string, password string) (result string, err error)
 	return
 }
 
-func RemoveUserPassword(user string, hashedPassword string) (result string, err error) {
+func RemoveUserPassword(user string, clearTextPassword string) (result string, err error) {
 	// Check for existing account
 	rows, err := Config.Db.Query("SELECT `accountNumber` FROM `accounts_auth` WHERE `accountNumber` = ?", user)
 	if err != nil {
@@ -148,6 +159,27 @@ func RemoveUserPassword(user string, hashedPassword string) (result string, err 
 		return "", errors.New("appauth.RemoveUserPassword: Account auth does not exists")
 	}
 
+	userHashedPassword, userSalt, err := getUserPasswordSaltFromUID(user)
+	if err != nil {
+		return "", errors.New("appauth.CreateUserPassword: Could not retrieve user details. " + err.Error())
+	}
+
+	// Generate hash
+	userPasswordSalt := userSalt + clearTextPassword
+	hashOutput, err := argon2.Key([]byte(userPasswordSalt), []byte(Config.PasswordSalt), 3, 4, 4096, 64, argon2.Argon2i)
+	if err != nil {
+		return "", errors.New("appauth.CreateUserPassword: Could not generate secure hash. " + err.Error())
+	}
+	hash := hex.EncodeToString(hashOutput)
+
+	fmt.Printf("clearTextPassword: %v\n", clearTextPassword)
+	fmt.Printf("userSalt: %v\n", userSalt)
+	fmt.Printf("userHashedPassword: %v\n", userHashedPassword)
+	fmt.Printf("generatedHash: %v\n", hash)
+	if hash != userHashedPassword {
+		return "", errors.New("appauth.CreateToken: Authentication credentials invalid")
+	}
+
 	// Prepare statement for inserting data
 	delStatement := "DELETE FROM accounts_auth WHERE `accountNumber` = ? AND `password` = ? "
 	stmtDel, err := Config.Db.Prepare(delStatement)
@@ -156,7 +188,16 @@ func RemoveUserPassword(user string, hashedPassword string) (result string, err 
 	}
 	defer stmtDel.Close() // Close the statement when we leave main() / the program terminates
 
-	_, err = stmtDel.Exec(user, hashedPassword)
+	res, err := stmtDel.Exec(user, userHashedPassword)
+
+	affected, err := res.RowsAffected()
+	if err != nil {
+		return "", errors.New("appauth.RemoveUserPassword: Could not get rows affected. " + err.Error())
+	}
+
+	if affected == 0 {
+		return "", errors.New("appauth.RemoveUserPassword: Could not delete account. No account deleted.")
+	}
 
 	if err != nil {
 		return "", errors.New("appauth.RemoveUserPassword: Could not delete account. " + err.Error())
@@ -167,7 +208,7 @@ func RemoveUserPassword(user string, hashedPassword string) (result string, err 
 }
 
 func CreateToken(user string, password string) (token string, err error) {
-	rows, err := Config.Db.Query("SELECT `password` FROM `accounts_auth` WHERE `accountNumber` = ?", user)
+	rows, err := Config.Db.Query("SELECT `password`, `salt` FROM `accounts_auth` WHERE `accountNumber` = ?", user)
 	if err != nil {
 		return "", errors.New("appauth.CreateToken: Error with select query. " + err.Error())
 	}
@@ -175,15 +216,17 @@ func CreateToken(user string, password string) (token string, err error) {
 
 	count := 0
 	hashedPassword := ""
+	userSalt := ""
 	for rows.Next() {
-		if err := rows.Scan(&hashedPassword); err != nil {
+		if err := rows.Scan(&hashedPassword, &userSalt); err != nil {
 			return "", errors.New("appauth.CreateToken: Could not retreive account details")
 		}
 		count++
 	}
 
 	// Generate hash
-	output, err := argon2.Key([]byte(password), []byte(Config.PasswordSalt), 3, 4, 4096, 32, argon2.Argon2i)
+	userPasswordSalt := userSalt + password
+	output, err := argon2.Key([]byte(userPasswordSalt), []byte(Config.PasswordSalt), 3, 4, 4096, 64, argon2.Argon2i)
 	if err != nil {
 		return "", errors.New("appauth.CreateUserPassword: Could not generate secure hash. " + err.Error())
 	}
@@ -253,6 +296,32 @@ func GetUserFromToken(token string) (user string, err error) {
 		if err != nil {
 			return "", errors.New("appauth.GetUserFromToken: Could not extend token. " + err.Error())
 		}
+	}
+
+	return
+}
+
+func RandStringBytes(n int) string {
+	b := make([]byte, n)
+	for i := range b {
+		b[i] = LETTER_BYTES[rand.Intn(len(LETTER_BYTES))]
+	}
+	return string(b)
+}
+
+func getUserPasswordSaltFromUID(user string) (hashedPassword string, userSalt string, err error) {
+	rows, err := Config.Db.Query("SELECT `password`, `salt` FROM `accounts_auth` WHERE `accountNumber` = ?", user)
+	if err != nil {
+		return "", "", errors.New("appauth.CreateToken: Error with select query. " + err.Error())
+	}
+	defer rows.Close()
+
+	count := 0
+	for rows.Next() {
+		if err := rows.Scan(&hashedPassword, &userSalt); err != nil {
+			return "", "", errors.New("appauth.CreateToken: Could not retreive account details")
+		}
+		count++
 	}
 
 	return
